requirements.txt
rich
python-telegram-bot